Second working GUI
This is the second draft working GUI for the preferences panel. It now updates mostly of userDefaults properties, with only a couple of them still in the work, which are related to the tableview array elements. It has a lot of console debug printing, just to make sure everything is working as expected.