Third working GUI
This is the third draft working GUI for the preferences panel. It now updates mostly of userDefaults properties, with only a couple of them still in the work. The TableView is being filled with the required elements, but still no editing of them available. It still has a lot of console debug printing, to make sure that everything is working as expected.
This patch adds some unused files cleanup, and a bunch of refactoring to better organize the code.