Start of general cleanup
This change cleans up the code and separate responsibilities of protocols and delegates by declaring them as extensions of the involved classes where they are required. It also isolates the console output messages to be present only on DEBUG builds. The general comments were also updated (sometimes completed/added) to follow other sources.
The arrays and dictionaries defined with constants inside classes definitions were rewritten to be more readable and consistent.
The Keys on PreferencesModel class were redefined to be used outside of the class, maintaining the class namespace bounding.
The value for the default Shortcut now is (command + shift + space).
Finally, the PreferencesModel class is now a subclass of NSObject, so it is now possible to begin the implementation of KVO for their properties.