Additional improvements
This change adds a custom class for the NSColorWell (AlphaColorWell) that adds the alpha selection for the standard NSColorWell.
There is a small change that fix the usage of some default constants in the PreferencesModel class definition.
It also fix a bug with the activationMode option, which could break the app is the preferences file became corrupted.