More updates to Preferences Panel
This change adds an extension to UserDefaults with the key identifiers for the preferences’ storage objects. It replaced the NSLog entries by debugPrint. Includes the KVO pattern for the MASShortcut control update. Remove the key identifiers from PreferencesModel, in anticipation of its future replacement by NSUserDefaultsController.
It also includes the workspace defined by CocoaPods.